% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{calc_var_e}
\alias{calc_var_e}
\title{Calculate Enviromental Varience on Assumed h2}
\usage{
calc_var_e(var_g, h2 = 0.3)
}
\arguments{
\item{h2}{}
}
\description{
Calculate Enviromental Varience on Assumed h2
}
\examples{
calc_var_e(0.1)
calc_var_e(0.1, 1) # should give 0 write test
}
