% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate Accuracy of Predictions}
\usage{
calculate_accuracy(ebv_table, percentage = 1, G1val = 1, G2val = 2)
}
\arguments{
\item{ebv_table}{}
}
\value{
accuracy for both tails
}
\description{
Calculate Accuracy of Predictions
}
