% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{calculate_accuracy_1}
\alias{calculate_accuracy_1}
\title{Calculate Accuracy for single tail of predictions}
\usage{
calculate_accuracy_1(ebv_table, percentage = 1, G1val = 1, G2val = 2)
}
\arguments{
\item{ebv_table}{}
}
\value{
accuracy for single tail
}
\description{
Calculate Accuracy for single tail of predictions
}
