% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{calculate_varience}
\alias{calculate_varience}
\title{Calculate Enviromental Variance of BV from h2}
\usage{
calculate_varience(bv, h2)
}
\arguments{
\item{bv}{vector containg breeding values}

\item{h2}{estimate of hertibility}
}
\value{
estimate of varience
}
\description{
Calculate Enviromental Variance of BV from h2
}
\examples{
calculate_varience(bv=rnorm(100),h2=0.5)
}
