% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{convert_geno_to_matrix}
\alias{convert_geno_to_matrix}
\title{Create a gentype matrix from the geno_list - NOT WORKING
Could potentially be simplified - needs test function}
\usage{
convert_geno_to_matrix(geno_list)
}
\arguments{
\item{geno_list}{a list of genotype matrices}
}
\value{
one matrix containing all genotypes
}
\description{
Create a gentype matrix from the geno_list - NOT WORKING
Could potentially be simplified - needs test function
}
\examples{
convert_geno_to_matrix(geno_list=list(rep(1,10)))
}
