% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{create_ebv_table}
\alias{create_ebv_table}
\title{Create EBV Table}
\usage{
create_ebv_table(ind_info, ebv)
}
\arguments{
\item{ind_info}{phenotypic assignment of test individuals}

\item{ebv}{estimated breeding value}
}
\value{
table of estimated breeding values and assigned groups for test individuals
}
\description{
Create EBV Table
}
