% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{create_ees_table}
\alias{create_ees_table}
\title{Write EES Results to File}
\usage{
create_ees_table(fit)
}
\arguments{
\item{fit}{model fit output to use}

\item{wd}{working directory}
}
\value{
a table printed to file ees.table in named working dirSNP	MAA.EES	MAA.EES.SE	MIA.EES	MIA.EES.SE
}
\description{
Write EES Results to File
}
\examples{
create_ees_table(mixed_solve_both(produce_sim_data(10, 100, 100)))
}
