% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_loop.R
\name{create_lm_table}
\alias{create_lm_table}
\title{Create a Data Table for GLM}
\usage{
create_lm_table(pool_rc_line, info)
}
\arguments{
\item{pool_rc_line}{a line from the pool-rc table + SNP  column!!}

\item{info}{table contain pop info}
}
\value{
a table for use in glm
}
\description{
Create a Data Table for GLM
}
\examples{
create_lm_table(cbind("SNP_N", fread("./extdata/example_pool_rc")), fread("./extdata/example_pop_data.csv"))
}
