% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{create_loci}
\alias{create_loci}
\title{Create Loci from Binomial Distribution}
\usage{
create_loci(snps, allelic_freq)
}
\arguments{
\item{snps}{number od snps being simulated}

\item{allelic_freq}{frequency of minor allele at each locus}
}
\value{
state of each locus
}
\description{
Create Loci from Binomial Distribution
}
\examples{
create_loci(100,rbeta(100, .1, .1))
}
