% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{exclude_group2}
\alias{exclude_group2}
\title{Exclude one named Group2 grouping
#can only exclude one at a time}
\usage{
exclude_group2(info, pool_data, group_name)
}
\arguments{
\item{group_name}{}
}
\description{
Exclude one named Group2 grouping
#can only exclude one at a time
}
