% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{find_pools_rc}
\alias{find_pools_rc}
\title{Find all .pools_rc files in a directory}
\usage{
find_pools_rc(pool_loc)
}
\arguments{
\item{pool_loc}{path to a directory containing wanted files}
}
\value{
list of pools_rc files
}
\description{
Find all .pools_rc files in a directory
}
\examples{
find_pools_rc("./extdata/Pools_RC")
}
