% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{find_top_snps}
\alias{find_top_snps}
\title{Pull Top GWAS Hits from pools.rc files}
\usage{
find_top_snps(pools_rc, hits, pool_info)
}
\arguments{
\item{pool_files}{list of pools_rc files - can be 1}

\item{top_hits}{table of snps to use}

\item{pool_info_file}{info on pools inc groupings and phenotype}
}
\value{
table with the top snps as listed in hits file
}
\description{
Pull Top GWAS Hits from pools.rc files
}
\examples{
find_top_snps(
fread("./extdata/Pools_RC/pools_rc.1"),
get_hits_from_file("./extdata/example_100_hits.gwas", 10),
fread("./extdata/example_pop_data.csv"))
}
