% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_functions.R
\name{find_varient_sites}
\alias{find_varient_sites}
\title{Indentify Variant and Invaraiant Sites}
\usage{
find_varient_sites(gt, MAF)
}
\arguments{
\item{gt}{matrix of genotypes for all populations}

\item{MAF}{minor allele frequency cutoff}
}
\value{
#boolean vector where TRUE is a varient site
}
\description{
Indentify Variant and Invaraiant Sites
}
\examples{
find_varient_sites(pool_sim(10, 100, 100, 0.5)$gt_matrix, 0.01)
}
