% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{fix_allele_mismatch}
\alias{fix_allele_mismatch}
\title{Find and Fix Major Allele Mismatches}
\usage{
fix_allele_mismatch(ees_table, gt, fix, major_snp)
}
\arguments{
\item{ees_table}{table of estimated effect sizes of snps}

\item{gt}{table of genotypes for indivduals 0/1/2}

\item{fix}{table of major/minor allele info for individuals}

\item{major_snp}{major snps in training population}
}
\value{
gt table corrected for any mismatched minor and major alleles
}
\description{
Find and Fix Major Allele Mismatches
}
\examples{
fix <- read_fix_table("./extdata/test.fix")
gt <- read_gt_table("./extdata/test.gt")
test_data <- read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), "./extdata/test.gwas", 10)
fix_allele_mismatch(fread("./extdata/test.ees_table"), gt, fix, test_data)
}
