% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{fraction_to_decimal}
\alias{fraction_to_decimal}
\title{Transform allele frequencies from fraction to decimal}
\usage{
fraction_to_decimal(fracData)
}
\arguments{
\item{fracData#}{a data frame containing only the fractions}
}
\value{
proportion of the alleles
}
\description{
Transform allele frequencies from fraction to decimal
}
