% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{generate_allelic_freqency}
\alias{generate_allelic_freqency}
\title{Generate Distribution of Allelic Frequencies for Minor Allele}
\usage{
generate_allelic_freqency(snps)
}
\arguments{
\item{snps}{total number of sites in simulation}
}
\value{
frequency of minor allele at each locus
}
\description{
Generate Distribution of Allelic Frequencies for Minor Allele
}
