% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{generate_effect_size}
\alias{generate_effect_size}
\title{Generate Distribution of Effect Sizes
To be used for training pools and test population}
\usage{
generate_effect_size(snps)
}
\arguments{
\item{snps}{total number of sites in simulation}
}
\value{
effect size of each locus
}
\description{
Generate Distribution of Effect Sizes
To be used for training pools and test population
}
