% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_training_pop.R
\name{generate_genotypes}
\alias{generate_genotypes}
\title{Simulate Genotype for a Group of Populations}
\usage{
generate_genotypes(num_pop, num_ind, num_sites, allelic_freq)
}
\arguments{
\item{num_pop}{number of populations}

\item{num_ind}{number of individuals in each population}

\item{num_sites}{number of sites under examination}

\item{allelic_freq}{frequency of minor allele at each site}
}
\value{
a list of genotype matrices
}
\description{
Simulate Genotype for a Group of Populations
}
\examples{
generate_genotypes(num_pop = 10, num_ind = 10, num_sites = 100, generate_allelic_freqency(100))
}
