% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{generate_population}
\alias{generate_population}
\title{Simulate SNPs for a Population}
\usage{
generate_population(ind, snps, allelic_freq)
}
\arguments{
\item{ind}{number of individuals in the population}

\item{snps}{number od snps being simulated}

\item{allelic_freq}{frequency of minor allele at each locus}
}
\value{
matrix containg loci states in a single population
}
\description{
Simulate SNPs for a Population
}
