% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{get_af_diff}
\alias{get_af_diff}
\title{Find Difference in Allele Frequency from Provinence Mean}
\usage{
get_af_diff(gt_matrix, prov_list)
}
\arguments{
\item{gt_matrix}{matrix of allele frequences}

\item{prov_list}{list of provinances}
}
\value{
matrix of deviation from prov means
}
\description{
Find Difference in Allele Frequency from Provinence Mean
}
