% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{get_allele}
\alias{get_allele}
\title{Get the Major/Minor Allele}
\usage{
get_allele(alleles, which)
}
\arguments{
\item{alleles}{column containing major minor e.g A/T}

\item{which}{which allele to get}
}
\value{
either major or minor allele
}
\description{
Get the Major/Minor Allele
}
\examples{
get_allele("A/T", "major")
}
