% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_loop.R
\name{get_allele_count}
\alias{get_allele_count}
\title{Get Allele Count}
\usage{
get_allele_count(allele_cols)
}
\arguments{
\item{allele_cols}{selected cols from pool_rc table}
}
\value{
the allele count from those cols
}
\description{
Get Allele Count
}
\examples{
get_allele_count(c("1/5", "3/7", "2/13"))
}
