% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{get_allele_freq}
\alias{get_allele_freq}
\title{Get the Major/Minor Allele Frequency}
\usage{
get_allele_freq(pool_rc, pool_info_file, which)
}
\arguments{
\item{pool_rc}{table of pools_rc data}

\item{pool_info_file}{file with info on training population}

\item{which}{which allele to get}
}
\value{
table with freqency of either major or minor alleles in each pool
}
\description{
Get the Major/Minor Allele Frequency
}
\examples{
get_allele_freq(fread("./extdata/example_pool_rc"), "./extdata/example_pop_data.csv", "minor"
}
