% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_training_pop.R
\name{get_breeding_values}
\alias{get_breeding_values}
\title{Calculate Breeding Values for a group of Populations}
\usage{
get_breeding_values(snp_list, effect_sizes)
}
\arguments{
\item{snp_list}{list of genotype matrices - one per population}

\item{effect_sizes}{a vector of effect sizes for each site}
}
\value{
list containing vectors of breeding values for each population
}
\description{
Calculate Breeding Values for a group of Populations
}
\examples{
pending!
}
