% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bv_estimation.R
\name{get_ebv}
\alias{get_ebv}
\title{Calculate EBV from GT and ES}
\usage{
get_ebv(ees_table, gt_matrix)
}
\arguments{
\item{ees_table}{data table containing results of rrblup}

\item{gt_matrix}{matrix of 0, 1, 2 and NA}
}
\value{
ebv vector
}
\description{
Calculate EBV from GT and ES
}
