% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{get_ees_subset}
\alias{get_ees_subset}
\title{Get a Subset of SNPs with Largest EES}
\usage{
get_ees_subset(ees_table, subset_size)
}
\arguments{
\item{ees_table}{table of estimated effect sizes of snps}

\item{subset_size}{size of subset to be taken from ees table}
}
\value{
subset of snps selected by ees
}
\description{
Get a Subset of SNPs with Largest EES
}
\examples{
get_ees_subset(fread("./extdata/test.ees_table"), 5)
}
