% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{get_exp_var}
\alias{get_exp_var}
\title{Calculate Expected Varience}
\usage{
get_exp_var(p, effect, exp_mean)
}
\arguments{
\item{exp_mean}{}
}
\description{
Calculate Expected Varience
}
\examples{
get_exp_var(0.2, 0.01, 0.1)
}
