% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_functions.R
\name{get_extreme_pheno}
\alias{get_extreme_pheno}
\title{Select the Individuals of Extreme Phenotype from each Population}
\usage{
get_extreme_pheno(pheno_list, i, cutoff)
}
\description{
Select the Individuals of Extreme Phenotype from each Population
}
