% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{get_hits_from_file}
\alias{get_hits_from_file}
\title{Get SNPs of Interest from GWAS Results}
\usage{
get_hits_from_file(p_val_file, num_hits)
}
\arguments{
\item{p_val_file}{file containing data: contig pos snp_id p-value}

\item{num_hits}{number of snps needed}
}
\value{
subset of input with lowest p-value
}
\description{
Get SNPs of Interest from GWAS Results
}
\examples{
get_hits_from_file("./extdata/example_100_hits.gwas", 10)
}
