% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{get_parameters}
\alias{get_parameters}
\title{Calaculate parameter matrix from a set of effect sizes}
\usage{
get_parameters(simData, effect)
}
\arguments{
\item{simData}{Simulated data object}

\item{effect}{Vector of effect sizes - from normal distribution}
}
\description{
Calaculate parameter matrix from a set of effect sizes
}
\examples{
data <- read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), "./extdata/test.gwas", 10)
get_parameters(data, 0.1)
sim_data <- produce_sim_data(10, 100, 100)
get_parameters(sim_data, 0.1)
}
