% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_training_pop.R
\name{get_pools}
\alias{get_pools}
\title{Get High and Low Pools}
\usage{
get_pools(sim, MAF = 0.01, threshold = 0.5)
}
\arguments{
\item{sim}{object containing simulated population}

\item{MAF}{minor allele frequency}

\item{threshold}{proportion of population to draw from each extreme
default set to 0.5}
}
\value{
list containing two matrices, hi and lo
}
\description{
Get High and Low Pools
}
\examples{
get_pools(sim_training_pops(1, 100, 100, 0.5), 0.01, 0.5)
get_pools(sim_training_pops(20, 1000, 1000, 0.2), 0.01, 0.2)
}
