% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{get_random}
\alias{get_random}
\title{Get SNPs of Interest from GWAS Results}
\usage{
get_random(pools_rc, num_hits)
}
\arguments{
\item{pools_rc}{data frame}

\item{num_hits}{number of snps needed}
}
\value{
random subset of rows from pools_rc
}
\description{
Get SNPs of Interest from GWAS Results
}
\examples{
get_random(fread("./extdata/test.pool_rc"), 10)
}
