% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_training_pop.R
\name{get_samples}
\alias{get_samples}
\title{Samples from Population}
\usage{
get_samples(pheno, threshold = 0.5)
}
\arguments{
\item{pheno}{matrix of dim 1*n where n is the number of individuals
in a population}

\item{threshold}{the proportion of the population taken from each extreme
default set to 0.5 (50:50) split}
}
\value{
two boolean vectors identifying which individuals
to use in high and low pools
}
\description{
Samples from Population
}
