% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{get_snp_id}
\alias{get_snp_id}
\title{Get Joined SNP ID from pools_rc format}
\usage{
get_snp_id(pool_rc)
}
\arguments{
\item{pool_rc}{file}
}
\value{
list of snps
}
\description{
Get Joined SNP ID from pools_rc format
}
\examples{
get_snp_id(fread("./extdata/example_pool_rc"))
}
