% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_loop.R
\name{get_snp_pvals}
\alias{get_snp_pvals}
\title{Get P values for each SNP}
\usage{
get_snp_pvals(pools_rc, pop_info)
}
\arguments{
\item{pools_rc}{a table containing read counts in each pool for snps of interest}

\item{pop_info_file}{the location of the file containing pool info}
}
\value{
a table containing p_values for each snp
}
\description{
Get P values for each SNP
}
\examples{
get_snp_pvals(fread("./extdata/example_pool_rc"), fread("./extdata/example_pop_data.csv"))
}
