% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_true_bv}
\alias{get_true_bv}
\title{Get True Breeding Value}
\usage{
get_true_bv(loci_of_pop, true_es)
}
\arguments{
\item{loci_of_pop}{matrix containg loci states in a single population}

\item{true_es}{true effect size of each locus}
}
\value{
true breeding value of each individual in the population
}
\description{
Get True Breeding Value
}
