% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{gppool_demo}
\alias{gppool_demo}
\title{Demo of Pipeline using Simulated Data}
\usage{
gppool_demo(n_pop = 10, n_ind = 1000, n_site = 1000, h2 = 0.3,
  MAF = 0.01, threshold = 0.2)
}
\arguments{
\item{n_pop}{number of populations to simulate}

\item{n_ind}{number of individuals per population}

\item{n_site}{number of sites in simulated genome}

\item{h2}{heritability of simulated trait}

\item{MAF}{minor allele theshold}

\item{threshold}{threshold for including simulated individuals in the high or low pools}
}
\value{
no return - prints output to console
}
\description{
Demo of Pipeline using Simulated Data
}
\examples{
gppool_demo(20, 100)
}
