% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{make_gt_matrix}
\alias{make_gt_matrix}
\title{Create a gentype matrix from the geno_list - NOT WORKING!!!
Could potentially be simplified - needs test function}
\usage{
make_gt_matrix(geno_list)
}
\arguments{
\item{geno_list}{a list of genotype matrices}
}
\value{
one matrix containing all genotypes
}
\description{
Create a gentype matrix from the geno_list - NOT WORKING!!!
Could potentially be simplified - needs test function
}
\examples{
make_gt_matrix(pool_sim(5, 10, 10, 0.5)$gt_list)
}
