% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{match_and_subset}
\alias{match_and_subset}
\title{Match Test & Training, Correct & Subset}
\usage{
match_and_subset(ees_table, gt, fix, pool_data, subset_size)
}
\arguments{
\item{ees_table}{table of estimated effect sizes of snps}

\item{gt}{table of genotypes for indivduals 0/1/2}

\item{fix}{table of major/minor allele info for individuals}

\item{pool_data}{data on training pools}

\item{subset_size}{size of subset of snps from test population to be used to calculate ebv}
}
\value{
list containing subsets of the ees and gt tables
}
\description{
Match Test & Training, Correct & Subset
}
\examples{
gt <- read_gt_table("./extdata/test.gt")
fix <- read_fix_table("./extdata/test.fix")
test_data <- read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), "./extdata/test.gwas", 10)
match_and_subset(fread("./extdata/test.ees_table"), gt, fix, test_data, 5)
}
