% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{match_snps_in_ind}
\alias{match_snps_in_ind}
\title{Find SNPs Occuring in Training and Test}
\usage{
match_snps_in_ind(ees_table, gt_ind)
}
\arguments{
\item{gt_ind}{}
}
\description{
Find SNPs Occuring in Training and Test
}
