% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{mixed_solve_both}
\alias{mixed_solve_both}
\title{Mixed Solve For Major and Minor Alleles}
\usage{
mixed_solve_both(data)
}
\arguments{
\item{data}{}
}
\description{
Mixed Solve For Major and Minor Alleles
}
\examples{
mixed_solve_both(produce_sim_data(10, 100, 1000))
}
