% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{ml_both_af_diff_X}
\alias{ml_both_af_diff_X}
\title{Mixed Solve Without Prov Effects Major and Minor Alleles}
\usage{
ml_both_af_diff_X(data)
}
\arguments{
\item{data}{}
}
\description{
Mixed Solve Without Prov Effects Major and Minor Alleles
}
\examples{
mixed_solve_both_af_diff(produce_sim_data(10, 100, 1000))
}
