% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{order_by_ees}
\alias{order_by_ees}
\title{Sort EES Table by MIA^2}
\usage{
order_by_ees(ees_table)
}
\arguments{
\item{ees_table}{table of estimated effect sizes of snps}
}
\value{
snps ordered by estimated effect size
}
\description{
Sort EES Table by MIA^2
}
\examples{
order_by_ees(fread("./extdata/test.ees_table")) - test that top > bottom
}
