% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_training_pop.R
\name{pool_freq}
\alias{pool_freq}
\title{Get Pooled Allele Frequency}
\usage{
pool_freq(gt, sample, variant_sites)
}
\arguments{
\item{gt}{matrix}

\item{sample}{boolean list}

\item{variant_sites}{boolean vector}
}
\value{
mean frequency of the minor allele at each varient site
}
\description{
Get Pooled Allele Frequency
}
