% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{pool_sim}
\alias{pool_sim}
\title{Simulate the Training Populations}
\usage{
pool_sim(num_pop, num_ind, num_sites, h2)
}
\arguments{
\item{num_pop}{number of populations}

\item{num_ind}{number of individuals in each population}

\item{num_sites}{number of sites under examination}
}
\value{
a list containing effect size of each site, its allelic frequency and a list with
the genotype matrices, breeding values and phenotypes(continuous)
}
\description{
Simulate the Training Populations
}
\examples{
 pool_sim(10, 100, 100, 0.5)
}
