% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_training_pop.R
\name{produce_sim_data}
\alias{produce_sim_data}
\title{Produce a Simulation data object}
\usage{
produce_sim_data(num_pop, num_ind, num_sites, h2 = 0.5, MAF = 0.01,
  threshold = 0.5)
}
\arguments{
\item{threshold}{}
}
\description{
Produce a Simulation data object
}
