% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_ind_data.R
\name{read_fix_table}
\alias{read_fix_table}
\title{Read in the fix table}
\usage{
read_fix_table(wd)
}
\arguments{
\item{wd}{working directory containing fix.table}
}
\value{
an 8 col data frame:
SNP CROM POS ID REF ALT QUAL FILTER
}
\description{
Read in the fix table
}
\examples{
read_fix_table("./extdata")
}
