% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_ind_data.R
\name{read_gt_table}
\alias{read_gt_table}
\title{Read in the genotype table}
\usage{
read_gt_table(gt_file, suffix = "MEMq20.sorted.bam")
}
\arguments{
\item{gt}{file working directory containing genotype.table
#Need to be able to vary suffix}
}
\value{
a data frame with named rows
}
\description{
Read in the genotype table
}
\examples{
read_gt_table("./extdata/test.gt")
}
