% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{read_in_pools_rc}
\alias{read_in_pools_rc}
\title{Get Data Structure from pools_rc}
\usage{
read_in_pools_rc(pools_rc_file, info, gwas, hit_num)
}
\arguments{
\item{hit_num}{}
}
\description{
Get Data Structure from pools_rc
}
\examples{
read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), "./extdata/test.gwas", 10)
}
