% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_ind_data.R
\name{read_ind_info}
\alias{read_ind_info}
\title{Read in Individual Infomation}
\usage{
read_ind_info(ind_info_file)
}
\arguments{
\item{ind_info_file}{path to file}
}
\value{
dataframe with two columns ID Group
}
\description{
Read in Individual Infomation
}
\examples{
read_ind_info("./extdata/example_ind_info.csv")
}
