% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{read_pools_rc_random}
\alias{read_pools_rc_random}
\title{Get random subset pools_rc}
\usage{
read_pools_rc_random(pools_rc_file, info, hit_num)
}
\arguments{
\item{pools_rc_file}{input file pools_rc}

\item{info}{table containing data on the grouping of pools <- needs to be read in}

\item{hit_num}{number of snps to be used to calculate ees}

\item{gwas}{file containing snps and p-values from gwas}
}
\value{
list of a random subset of data read in from pools_rc
}
\description{
Get random subset pools_rc
}
\examples{
read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), 10)
}
