% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{remove_prov_effects}
\alias{remove_prov_effects}
\title{Find Difference from Provenance Mean Frequency}
\usage{
remove_prov_effects(dec, prov)
}
\arguments{
\item{prov}{}
}
\description{
Find Difference from Provenance Mean Frequency
}
