% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{rrblup_loop}
\alias{rrblup_loop}
\title{Run rrBLUP as loop}
\usage{
rrblup_loop(data, X = FALSE, Diff = TRUE, both = TRUE, rep = 10)
}
\arguments{
\item{data}{object containing data from training population}

\item{X}{Boolean - should attempt be made to fit X - default FALSE}

\item{Diff}{Should the difference from Group2 mean be used in place of raw values - default = TRUE}

\item{both}{Should both minor and msjor alleles be used - default = TRUE}

\item{rep}{How many times should model fitting be repeated}
}
\value{
list containing vectors with estimated effect sizes and snp_id
}
\description{
Run rrBLUP as loop
}
