% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_test_pop.R
\name{sim_test_gt}
\alias{sim_test_gt}
\title{Simulate gt matrix for test population}
\usage{
sim_test_gt(num_ind, af)
}
\arguments{
\item{num_ind}{number of individuals}

\item{af}{vector of allelic frequency}
}
\value{
matrix of genotypes: 0, 1, 2
}
\description{
Simulate gt matrix for test population
}
\examples{
sim_test_gt(100, generate_allelic_freqency(100))
}
