% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_test_pop.R
\name{sim_test_pop}
\alias{sim_test_pop}
\title{Simulate the test population}
\usage{
sim_test_pop(es, af, h2 = 0.3, test_ind = 100)
}
\arguments{
\item{es}{vector containing estimated effect sizes - eg from training pop}

\item{af}{vector containing allelic frequency - eg from training pop}

\item{h2}{heritability}

\item{test_ind}{number of individuals in test population}
}
\description{
Simulate the test population
}
\examples{
train <- sim_training_pops(10, 100, 100, 0.5)
sim_test_pop(train$es, train$af)
}
