% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_training_pop.R
\name{sim_training_pops}
\alias{sim_training_pops}
\title{Simulate the Training Populations}
\usage{
sim_training_pops(num_pop, num_ind, num_sites, h2)
}
\arguments{
\item{num_pop}{number of populations}

\item{num_ind}{number of individuals in each population}

\item{num_sites}{number of sites under examination}
}
\value{
a list containing effect size of each site, its allelic frequency, the genotype matrix and a list with
vectors of breeding values and phenotypes(continuous)
}
\description{
Simulate the Training Populations
}
\examples{
 sim_training_pops(10, 100, 100, 0.5)
}
