% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_loop.R
\name{split_allele_count}
\alias{split_allele_count}
\title{Split Allele count Cell}
\usage{
split_allele_count(allele_count)
}
\arguments{
\item{allele_count}{alele count in format "1/2}
}
\value{
allele count
}
\description{
Split Allele count Cell
}
\examples{
split_allele_count("10/20")
}
