% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrblup.R
\name{write_ees_table}
\alias{write_ees_table}
\title{Write EES Results to File}
\usage{
write_ees_table(wd, fit)
}
\arguments{
\item{wd}{working directory}

\item{fit}{model fit output to use}
}
\value{
a table printed to file ees.table in named working dirSNP	MAA.EES	MAA.EES.SE	MIA.EES	MIA.EES.SE
}
\description{
Write EES Results to File
}
\examples{
write_ees_table("./extdata", mixed_solve_both(produce_sim_data(10, 100, 100)))
}
