% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate Correlation of EBV and Group}
\usage{
calculate_correlation(ebv_table)
}
\arguments{
\item{ebv_table}{}
}
\value{
correlation
}
\description{
Calculate Correlation of EBV and Group
}
