% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{exclude_group2}
\alias{exclude_group2}
\title{Exclude one named Group2 grouping
#can only exclude one at a time}
\usage{
exclude_group2(info, pool_data, group_name)
}
\arguments{
\item{info}{data table with 3 columns}

\item{pool_data}{data output by read_in_pools_rc}

\item{group_name}{a group in column 3 (Group2) of the info table}
}
\value{
pool_data excluding the named group
}
\description{
Exclude one named Group2 grouping
#can only exclude one at a time
}
