% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{find_top_snps}
\alias{find_top_snps}
\title{Pull Top GWAS Hits from pools.rc files}
\usage{
find_top_snps(pool_files, hits, pool_col_names)
}
\arguments{
\item{pool_col_names}{}
}
\value{
table with the top snps as listed in hits file
}
\description{
Pull Top GWAS Hits from pools.rc files
}
