% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{fix_allele_mismatch}
\alias{fix_allele_mismatch}
\title{Find and Fix Major Allele Mismatches}
\usage{
fix_allele_mismatch(ees_table, gt_table, fix_table, pop_data)
}
\arguments{
\item{pop_data}{}
}
\description{
Find and Fix Major Allele Mismatches
}
