% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_training_pop.R
\name{generate_allelic_freqency}
\alias{generate_allelic_freqency}
\title{Generate Distribution of Allelic Frequencies for Minor Allele}
\usage{
generate_allelic_freqency(num_sites)
}
\arguments{
\item{num_sites}{total number of num_sites in simulation}
}
\value{
frequency of minor allele at each locus
}
\description{
Generate Distribution of Allelic Frequencies for Minor Allele
}
\examples{
generate_allelic_freqency(1000)
}
