% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_training_pop.R
\name{generate_effect_size}
\alias{generate_effect_size}
\title{Generate Distribution of Effect Sizes
To be used for training pools and test population}
\usage{
generate_effect_size(num_sites)
}
\arguments{
\item{num_sites}{total number of num_sites in simulation}
}
\value{
effect size of each locus
}
\description{
Generate Distribution of Effect Sizes
To be used for training pools and test population
}
\examples{
generate_effect_size(1000)
}
