% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_training_pop.R
\name{generate_population}
\alias{generate_population}
\title{Simulate num_sites for a Population}
\usage{
generate_population(ind, num_sites, allelic_freq)
}
\arguments{
\item{ind}{number of individuals in the population}

\item{num_sites}{number od num_sites being simulated}

\item{allelic_freq}{frequency of minor allele at each locus}
}
\value{
matrix containg loci states in a single population
}
\description{
Simulate num_sites for a Population
}
\examples{
generate_population(10,100, generate_allelic_freqency(100))
}
