% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{get_ees_subset}
\alias{get_ees_subset}
\title{Get a Subset of SNPs with Largest EES}
\usage{
get_ees_subset(ees_table, subset_size)
}
\arguments{
\item{subset_size}{}
}
\description{
Get a Subset of SNPs with Largest EES
}
