% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{get_gt_subset}
\alias{get_gt_subset}
\title{Subset GT Matrix by SNP list}
\usage{
get_gt_subset(snp_list, gt)
}
\arguments{
\item{gt}{}
}
\description{
Subset GT Matrix by SNP list
}
\examples{
ind_gt <- read_gt_table("./extdata/test.gt")
get_gt_subset(get_ees_subset(fread("./extdata/test.ees_table"), 5)$SNP, ind_gt)
}
