% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_phenotype}
\alias{get_phenotype}
\title{Get Phenotypic Value}
\usage{
get_phenotype(bv, est_var)
}
\arguments{
\item{bv}{vector containg breeding values}

\item{est_var}{estimate of varience}
}
\value{
vector containing phenotypes
}
\description{
Get Phenotypic Value
}
