% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{gppool_data_demo}
\alias{gppool_data_demo}
\title{Demo using example dataset}
\usage{
gppool_data_demo(training_snps = 100, test_snps = 50)
}
\arguments{
\item{test_snps}{}
}
\value{
print output to terminal and create boxplot
}
\description{
Demo using example dataset
}
