% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_training_and_test.R
\name{match_and_subset}
\alias{match_and_subset}
\title{Match Test & Training, Correct & Subset}
\usage{
match_and_subset(ees_table, gt, fix, pool_data, subset_size)
}
\arguments{
\item{subset_size}{}
}
\description{
Match Test & Training, Correct & Subset
}
\examples{
ind_gt <- read_gt_table("./extdata/test.gt")
match_and_subset(fread("./extdata/test.ees_table"), ind_gt, ind_fix, test_data, 5)
}
