% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{plot_ebv_boxplot}
\alias{plot_ebv_boxplot}
\title{Plot Boxplot of EBV (ggplot)}
\usage{
plot_ebv_boxplot(ebv_table)
}
\arguments{
\item{ebv_table}{table of estimated breeding values and assigned groups for test individuals}
}
\value{
boxlot
}
\description{
Plot Boxplot of EBV (ggplot)
}
