% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{read_in_pools_rc}
\alias{read_in_pools_rc}
\title{Get Data Structure from pools_rc}
\usage{
read_in_pools_rc(pools_rc_file, info, gwas, hit_num)
}
\arguments{
\item{pools_rc_file}{input file pools_rc}

\item{info}{table containing data on the grouping of pools <- needs to be read in}

\item{gwas}{file containing snps and p-values from gwas}

\item{hit_num}{number of snps to be used to calculate ees}
}
\value{
list data read in from pools_rc
}
\description{
Get Data Structure from pools_rc
}
\examples{
read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), "./extdata/test.gwas", 10)
}
