% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pools_rc.R
\name{read_pools_rc_random}
\alias{read_pools_rc_random}
\title{Get random subset pools_rc}
\usage{
read_pools_rc_random(pools_rc_file, info, hit_num)
}
\arguments{
\item{hit_num}{}
}
\description{
Get random subset pools_rc
}
\examples{
read_in_pools_rc("./extdata/test.pool_rc", fread("./extdata/test.pool_info"), 10)
}
